Upgrade required conan version
A recent upstream change to the `googleapis` conan recipe broke the
conan build for some folks in some situation with an error message
saying that protobuf was not found.

I analysed the problem and it turns out that the update of `googleapis`
triggers a known and fixed bug in Conan and an upgrade of Conan fixes
the problem.