Show Top Down Tab after capture
For historical reasons, we showed the sampling tab when completing
a capture and no instrumation was done.

This behaviour changed recently, such that we were not switching
the tab in this case at all, and remain at the modules tab.
However, it makes sense to inspect sampling related data after
the capture is done.
This change suggest to switch to the top down tab, as most other
profilers do.

Test: Take a capture w/ and w/o instrumentation.