Fix drawing of incomplete intervals (#4842)
Incomplete intervals, drawn as orange translucid rectangles, could be drawn over
the track headers on the left side of the capture view. Make sure we clamp
the x coordinate. Tested by introducing fake intervals and making sure they aligned
properly on the timeline and that they did not draw over the track headers.